In case of SIGINT signal, throw SigIntException

+++ src/lib/syscalls/SysCalls.h
@@ -1,6 +1,7 @@
 #ifndef SRC_UTILS_SYSCALLS_H_
 #define SRC_UTILS_SYSCALLS_H_
 
+#include <asm-generic/errno-base.h>
 #include <fcntl.h>
 #include <stddef.h>
 #include <sys/ipc.h>
@@ -8,11 +9,13 @@
 #include <sys/shm.h>
 #include <sys/stat.h>
 #include <unistd.h>
+#include <cerrno>
 #include <cstdio>
 #include <string>
 #include <vector>
 
 #include "exceptions/SysCallException.h"
+#include "exceptions/SigIntException.h"
 
 using namespace std;
 
@@ -23,6 +26,9 @@ inline pid_t fork() {
 	pid_t id = ::fork();
 
 	if (id < 0) {
+		if (errno == EINTR){
+			throw SigIntException();
+		}
 		throw SysCallException("Error trying to execute a fork()");
 	}
 
@@ -35,6 +41,9 @@ inline pid_t getpid() {
 	pid_t result = ::getpid();
 
 	if (result < 0) {
+		if (errno == EINTR){
+			throw SigIntException();
+		}
 		throw SysCallException("Error trying to execute a getpid()");
 	}
 
@@ -54,6 +63,9 @@ inline void execv(const string &execPath, vector<string> params) {
 	perror("execError");
 
 	if (status < 0) {
+		if (errno == EINTR){
+			throw SigIntException();
+		}
 		throw SysCallException("Error trying to execute a execv()");
 	}
 }
@@ -61,6 +73,9 @@ inline void execv(const string &execPath, vector<string> params) {
 inline int mknod(const char *pathname, mode_t mode, dev_t dev) {
 	int response = ::mknod(pathname, mode, dev);
 	if (response < 0) {
+		if (errno == EINTR){
+			throw SigIntException();
+		}
 		throw SysCallException("Error trying to execute a mknod()");
 	}
 	return response;
@@ -69,6 +84,9 @@ inline int mknod(const char *pathname, mode_t mode, dev_t dev) {
 inline int unlink(const char *pathname) {
 	int response = ::unlink(pathname);
 	if (response < 0) {
+		if (errno == EINTR){
+			throw SigIntException();
+		}
 		throw SysCallException("Error trying to execute a unlink()");
 	}
 	return response;
@@ -77,6 +95,9 @@ inline int unlink(const char *pathname) {
 inline int close(int fd) {
 	int response = ::close(fd);
 	if (response < 0) {
+		if (errno == EINTR){
+			throw SigIntException();
+		}
 		throw SysCallException("Error trying to execute a close()");
 	}
 	return response;
@@ -85,6 +106,9 @@ inline int close(int fd) {
 inline int open(const char *pathname, int flags, int otherFlags) {
 	int response = ::open(pathname, flags, otherFlags);
 	if (response < 0) {
+		if (errno == EINTR){
+			throw SigIntException();
+		}
 		throw SysCallException("Error trying to execute a open()");
 	}
 	return response;
@@ -93,6 +117,9 @@ inline int open(const char *pathname, int flags, int otherFlags) {
 inline int open(const char *pathname, int flags) {
 	int response = ::open(pathname, flags);
 	if (response < 0) {
+		if (errno == EINTR){
+			throw SigIntException();
+		}
 		throw SysCallException("Error trying to execute a open()");
 	}
 	return response;
@@ -101,6 +128,9 @@ inline int open(const char *pathname, int flags) {
 inline ssize_t read(int fd, void *buf, size_t count) {
 	int response = ::read(fd, buf, count);
 	if (response < 0) {
+		if (errno == EINTR){
+			throw SigIntException();
+		}
 		throw SysCallException("Error trying to execute a read()");
 	}
 	return response;
@@ -109,6 +139,9 @@ inline ssize_t read(int fd, void *buf, size_t count) {
 inline ssize_t write(int fd, const void *buf, size_t count) {
 	int response = ::write(fd, buf, count);
 	if (response < 0) {
+		if (errno == EINTR){
+			throw SigIntException();
+		}
 		throw SysCallException("Error trying to execute a write()");
 	}
 	return response;
@@ -117,6 +150,9 @@ inline ssize_t write(int fd, const void *buf, size_t count) {
 inline off_t lseek(int fd, off_t offset, int whence) {
 	off_t response = ::lseek(fd, offset, whence);
 	if (response < 0) {
+		if (errno == EINTR){
+			throw SigIntException();
+		}
 		throw SysCallException("Error trying to execute a lseek()");
 	}
 	return response;
@@ -125,6 +161,9 @@ inline off_t lseek(int fd, off_t offset, int whence) {
 inline key_t ftok(const char *pathname, int proj_id) {
 	key_t response = ::ftok(pathname, proj_id);
 	if (response < 0) {
+		if (errno == EINTR){
+			throw SigIntException();
+		}
 		throw SysCallException("Error trying to execute a ftok()");
 	}
 	return response;
@@ -133,6 +172,9 @@ inline key_t ftok(const char *pathname, int proj_id) {
 inline int semget(key_t key, int nsems, int semflg) {
 	int response = ::semget(key, nsems, semflg);
 	if (response < 0) {
+		if (errno == EINTR){
+			throw SigIntException();
+		}
 		throw SysCallException("Error trying to execute a semget()");
 	}
 	return response;
@@ -141,6 +183,9 @@ inline int semget(key_t key, int nsems, int semflg) {
 inline int semop(int semid, struct sembuf *sops, size_t nsops) {
 	int response = ::semop(semid, sops, nsops);
 	if (response < 0) {
+		if (errno == EINTR){
+			throw SigIntException();
+		}
 		throw SysCallException("Error trying to execute a semop()");
 	}
 	return response;
@@ -149,6 +194,9 @@ inline int semop(int semid, struct sembuf *sops, size_t nsops) {
 inline int shmget(key_t key, size_t size, int shmflg) {
 	int response = ::shmget(key, size, shmflg);
 	if (response < 0) {
+		if (errno == EINTR){
+			throw SigIntException();
+		}
 		throw SysCallException("Error trying to execute a shmget()");
 	}
 	return response;
@@ -157,6 +205,9 @@ inline int shmget(key_t key, size_t size, int shmflg) {
 inline void *shmat(int shmid, const void *shmaddr, int shmflg) {
 	void* response = ::shmat(shmid, shmaddr, shmflg);
 	if (response < 0) {
+		if (errno == EINTR){
+			throw SigIntException();
+		}
 		throw SysCallException("Error trying to execute a shmat()");
 	}
 	return response;
@@ -165,6 +216,9 @@ inline void *shmat(int shmid, const void *shmaddr, int shmflg) {
 inline int semctl(int semid, int senum, int cmd) {
 	int response = ::semctl(semid, senum, cmd);
 	if (response < 0) {
+		if (errno == EINTR){
+			throw SigIntException();
+		}
 		throw SysCallException("Error trying to execute a semctl()");
 	}
 	return response;
@@ -173,6 +227,9 @@ inline int semctl(int semid, int senum, int cmd) {
 inline int shmdt(const void *shmaddr) {
 	int response = ::shmdt(shmaddr);
 	if (response < 0) {
+		if (errno == EINTR){
+			throw SigIntException();
+		}
 		throw SysCallException("Error trying to execute a shmdt()");
 	}
 	return response;
@@ -181,6 +238,9 @@ inline int shmdt(const void *shmaddr) {
 inline int shmctl(int shmid, int cmd, struct shmid_ds *buf) {
 	int response = ::shmctl(shmid, cmd, buf);
 	if (response < 0) {
+		if (errno == EINTR){
+			throw SigIntException();
+		}
 		throw SysCallException("Error trying to execute a shmctl()");
 	}
 	return response;
